Modulo 7/Clase7-8/App/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
streamlit
psycopg2-binary